Collapse single volumes and issues
That is, if a series only has one volume, treat it as though it doesn't actually use volumes. If a volume only has one issue, treat it as though it doesn't actually use issues. This lets creators have comic series that appear to simply be composed of strips, without any method of breaking them down into smaller parts.